requirements.txt
aniso8601==7.0.0
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
astroid==2.4.2
certifi==2020.6.20
colorama==0.4.3
coverage==5.3
Django==3.0.3
django-cors-headers==3.5.0
djangorestframework==3.11.0
graphene==2.1.8
graphene-django==2.12.1
graphql-core==2.3.2
graphql-relay==2.0.1
isort==5.5.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
promise==2.3
pylint==2.6.0
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Unidecode==1.1.1
wincertstore==0.2
wrapt==1.12.1
